start of final game creation
